Moves new script files to scripts
