+++ src/main/webapp/scripts/Attendee.js
