+++ src/main/webapp/scripts/Attendee.test.js
