+++ src/main/webapp/scripts/Session.js
